# Fix Hive applyProjection performance issue