@findepi 2020-11-02
I suggest to get `Build variable mapping set only once outside of the stream` in (already tested and reviewed here).

For `Ignore type in equals and hashCode for ConnectorExpression` i think we should rather fix type equality (https://github.com/prestosql/presto/pull/5768#discussion_r515552087)
and precalculating hashCode (https://github.com/prestosql/presto/pull/5768#discussion_r515558215) seems reasonable to me.