@phd3 2020-11-02
@findepi that sounds good, updated the PR. 

The issue described in the slack thread is ameliorated with this PR in the sense that the optimizer will not exhaust the limit.  However, the cpu usage of `applyProjection` is still unreasonably high because of the `ConnectorExpression` set building cost. 

The [explain query](https://github.com/prestosql/presto/files/5476020/query.txt) with ~1500 output dereference didn't finish before. With the commit in this PR, it now finishes in 8s, which is still too long. I haven't included it as a test, since this seems like a very trivial change.

Tracking the issue around type equality and hashing in #5780. I think that's also a critical one, so let me try to open up a PR soon and hopefully get in the next release.