@findepi 2020-11-03
> The [explain query](https://github.com/prestosql/presto/files/5476020/query.txt) with ~1500 output dereference didn't finish before. With the commit in this PR, it now finishes in 8s, which is still too long. I haven't included it as a test, since this seems like a very trivial change.

@phd3  BTW we have `BenchmarkPlanner`.
I would be surprised if it actually covered the problem being solved here, because there are no ROW types involved in TPCH queries used in that benchmark.
Maybe it would be worthwile to extend the benchmark after all? 
To cover the fix here, https://github.com/prestosql/presto/issues/5780 and any other improvement in that space?